generator_ng/aws_requirements.txt
boto3==1.17.75
botocore==1.20.75
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.9
idna==2.10
jmespath==0.10.0
py==1.10.0
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
retry==0.9.2
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
xlsxwriter==3.0.1
pandas==1.3.2
openpyxl==3.0.7
